+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Next
 
+* Added Activatedness information to Set Bonuses on armor held by a character.
+* Builds in Loadout Optimizer are now sorted by enabled stats, then each enabled stat in order, then total stats (including disabled stats). Before, they were not sorted by total stats, so if you had some stats disabled you could get very low-stat builds near the top.
+
 ## 8.85.0 <span class="changelog-date">(2025-08-10)</span>
 
 * Improved the responsiveness of draggable lists (Loadout Optimizer stat constraints, inventory sort order, etc.)